backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
numpy==1.26.2
scipy==1.11.4
numpy-stl==3.1.1
weasyprint==60.1
pandas==2.1.3
openpyxl==3.1.2
aiosqlite==0.19.0
python-jose==3.3.0
passlib==1.7.4
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
